refactor: restructure multimodal module
