+++ lib/mobility-core/src/Kernel/External/MultiModal/OpenTripPlanner/query.gql
