+++ lib/mobility-core/src/Kernel/External/MultiModal/OpenTripPlanner/schema.gql
